made files norminette-compliant and fixed ft_itoa_base handling of min int

+++ ft_itoa_base.c
@@ -6,7 +6,7 @@
 /*   By: bteuling <bteuling@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/06/17 17:25:33 by bteuling      #+#    #+#                 */
-/*   Updated: 2022/06/20 15:47:46 by bteuling      ########   odam.nl         */
+/*   Updated: 2022/06/20 16:41:29 by bteuling      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,11 +21,11 @@
  * @param base_string alphanumeric string which is the base used to convert n
  * @return char
  */
-char	*ft_itoa_base(int n, int base, char *base_string)
+char	*ft_itoa_base(long long n, int base, char *base_string)
 {
 	char	*converted_n;
 	int		len;
-	long num;
+	long	num;
 
 	len = ft_intlen_base(n, base);
 	converted_n = malloc((len + 1) * sizeof(char));
